pyproject.toml
[project]
name = "hexlet-code"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "flask==3.0.0",
    "python-dotenv==1.0.0",
    "gunicorn==21.2.0",
    "psycopg2-binary==2.9.9",
    "validators==0.22.0",
    "requests==2.31.0",
    "beautifulsoup4==4.12.2",
    "python-dateutil==2.8.2"
]

[build-system]
requires = ["setuptools>=64.0.0"]
build-backend = "setuptools.build_meta"